Split delimit_input func to new files
