+++ Makefile
@@ -22,6 +22,7 @@ LIBREADLINE_DIR	= readline
 # Files
 SRC				= $(SRC_DIR)/main.c \
 					$(SRC_DIR)/lexer/constructor.c \
+					$(SRC_DIR)/lexer/delimit_input.c \
 					$(SRC_DIR)/lexer/get_next_token.c \
 					$(SRC_DIR)/lexer/is_start_of_input.c \
 					$(SRC_DIR)/lexer/lexer.c \
